Setting up dash, docker etc
